global variable removed from db_mmap_sa
git-svn-id: http://svn.miranda-ng.org/main/trunk@3055 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/Db3x_mmap/src/dbsettings.cpp
@@ -328,7 +328,7 @@ STDMETHODIMP_(BOOL) CDb3Base::FreeVariant(DBVARIANT *dbv)
 
 STDMETHODIMP_(BOOL) CDb3Base::SetSettingResident(BOOL bIsResident, const char *pszSettingName)
 {
-	char *szSetting = m_cache->GetCachedSetting(NULL, pszSettingName, 0, strlen(pszSettingName));
+	char *szSetting = m_cache->GetCachedSetting(NULL, pszSettingName, 0, (int)strlen(pszSettingName));
 	szSetting[-1] = (char)bIsResident;
 
 	mir_cslock lck(m_csDbAccess);
